Roster: roster vector now decimal[] KP-2934
Former-commit-id: 9b4700b3003f4963293d221a48937fa1d45dd564

+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Events/Interview/AnswerCommented.cs
@@ -9,7 +9,7 @@ namespace WB.Core.SharedKernels.DataCollection.Events.Interview
 
         public string Comment { get; private set; }
 
-        public AnswerCommented(Guid userId, Guid questionId, int[] propagationVector, DateTime commentTime, string comment)
+        public AnswerCommented(Guid userId, Guid questionId, decimal[] propagationVector, DateTime commentTime, string comment)
             : base(userId, questionId, propagationVector)
         {
             this.Comment = comment;
