+++ src/Tests/WB.Core.SharedKernels.DataCollection.Tests/InterviewTests/when_answering_mandatory_unanswered_question.cs
@@ -52,7 +52,7 @@ namespace WB.Core.SharedKernels.DataCollection.Tests.InterviewTests
         };
 
         Because of = () =>
-            interview.AnswerNumericRealQuestion(userId, mandatoryQuestionId, new int[] { }, DateTime.Now, 0);
+            interview.AnswerNumericRealQuestion(userId, mandatoryQuestionId, new decimal[] { }, DateTime.Now, 0);
 
         It should_raise_AnswerDeclaredValid_event_with_QuestionId_equal_to_mandatoryQuestionId = () =>
             eventContext.ShouldContainEvent<AnswerDeclaredValid>(@event
